take context back out of the equation
it will be put back in as a variable member of the Resource trait. This will be safe because webmachine runs through a resource on a single thread. ReqRespData could also be treated this way but then it wouldn't be accessible outside to the decisions without exposing it as a public member which would be narsty. This also means that context can be ignored and customized using class members or other alternatives. On the downside it means that resources cannot be defined as modules like the original implementation, however, it will provide a more type safe context so I find this trade-off worth it.